pyproject.toml
[project]
name = "sse-relay-server"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Tobi DEGNON", email = "tobidegnon@proton.me" }
]
dependencies = [
    "starlette>=0.31.1",
    "sse-starlette>=1.6.5",
    "psycopg[c]>=3.1.10",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.23.2",
    "dj_database_url>=2.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.9.1",
]

[tool.rye.scripts]
sse-relay-server = { call = "sse_relay_server.main" }


[tool.hatch.metadata]
allow-direct-references = true
